refactor: internal test config
